% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epsg_table}
\alias{epsg_table}
\title{epsg_table}
\format{A SpatialPolygonsDataframe object with many
\describe{
  \item{name}{Name of the island}
  ...
}}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
epsg_table
}
\description{
A dataset of epsg codes and corresponding strings that can be used
to create a CRS object.
}
\keyword{datasets}
