% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_CRS.R
\name{find_CRS}
\alias{find_CRS}
\title{This function allows you to go from shp to df, and specify a projection}
\usage{
find_CRS(epsg = NULL, search = NULL)
}
\arguments{
\item{epsg}{the epsg name - in character}

\item{search}{Search term - e.g. "New Zealand"}
}
\description{
This function allows you to go from shp to df, and specify a projection
}
\examples{
# Find crs for NZTM, which is epsg 2193
find_CRS(epsg = "2193")
}
\keyword{epsg}
