% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_my_island.R
\name{find_my_island}
\alias{find_my_island}
\title{Getting islands in high res & correct long lats}
\usage{
find_my_island(islandname, proj = "wgs84")
}
\arguments{
\item{islandname}{The name of the island you are looking for}

\item{proj}{The name of the projection you would like the shapefile returned inCurrently either only "wgs84" or "nztm"}
}
\description{
This function allows you to go from shp to df, and specify a projection
}
\examples{
isle <- find_my_island("codfish", proj = "nztm")
plot(isle)
isle@proj4string
# for wgs84
isle <- find_my_island("codfish", proj = "wgs84")
plot(isle)
isle@proj4string
# multiple islands
many <- find_my_island(c("auckland", "campbell"))
plot(many)
}
\keyword{map}
\keyword{reproject}
\keyword{transform}
