% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortifying_polygons.R
\name{fortify_polygons}
\alias{fortify_polygons}
\title{Converting spatial polygon dataframes to dfs}
\usage{
fortify_polygons(shape, proj)
}
\arguments{
\item{shape}{The name of your shapefile object}

\item{proj}{The name of the projection. Currently either only "wgs84" or "nztm"}
}
\description{
This function allows you to go from shp to df, and specify a projection
}
\keyword{reproject}
\keyword{transform}
