% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzHigh}
\alias{nzHigh}
\title{nzHigh}
\format{
A SpatialPolygonsDataframe object with 3 features
\describe{
  \item{name}{Name of the island}
  ...
}
}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
nzHigh
}
\description{
A SpatialPolygonsDataframe of New Zealand in high resolution and nztm projection
}
\keyword{datasets}
