% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzMed}
\alias{nzMed}
\title{nzMed}
\format{A SpatialPolygonsDataframe object with 3 features
\describe{
  \item{name}{Name of the island}
  ...
}}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
nzMed
}
\description{
A SpatialPolygonsDataframe of New Zealand in medium resolution and nztm projection
}
\keyword{datasets}
