% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzSml}
\alias{nzSml}
\title{nzSml}
\format{A SpatialPolygonsDataframe object with 3 features
\describe{
  \item{name}{Name of the island}
  ...
}}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
nzSml
}
\description{
A SpatialPolygonsDataframe of New Zealand in low resolution and nztm projection
}
\keyword{datasets}
