% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzSml84}
\alias{nzSml84}
\title{nzSml84}
\format{
A SpatialPolygonsDataframe object with 3 features
\describe{
  \item{name}{Name of the island}
  ...
}
}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
nzSml84
}
\description{
A SpatialPolygonsDataframe of New Zealand in low resolution and wgs84 projection
}
\keyword{datasets}
