% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_in_NZTM.R
\name{readNZTM}
\alias{readNZTM}
\title{Converting spatial polygon dataframes to dfs}
\usage{
readNZTM(dsn, layer)
}
\arguments{
\item{dsn}{The path from current working directory to the folder containing the shapefile}

\item{layer}{The name of the shapefile without the .shp}
}
\description{
This function allows you to go from shp to df, and specify a projection
}
\keyword{import}
