% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_df.R
\name{spatial_df}
\alias{spatial_df}
\title{spatial_df}
\usage{
spatial_df(data, xcol, ycol, crs)
}
\arguments{
\item{data}{dataframe with coords and other data in}

\item{xcol}{name of column with x coords in}

\item{ycol}{name of column with y coords in}

\item{crs}{CRS object. Can be from find_CRS}
}
\value{
spatialpointsdataframe
}
\description{
spatial_df
}
\examples{
tp <- data.frame(
Island = c("South Island", "North Island",
"North Island", "Stewart Island", "South Island"),
City = c("Christchurch", "Wellington", "Auckland", "Oban", "Dunedin"),
long = c(172.63333,174.77557, 174.76667, 168.12736, 170.50361),
lat = c(-43.53333, -41.28664, -36.86667, -46.89881, -45.87416))

wgs84 <- CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs ")
# we know nztm isespg2193
nztm <- find_CRS(epsg = "2193")

spatialCities <- spatial_df(data = tp, xcol = "long", ycol = "lat",
crs = wgs84)

cols <- c("forestgreen", "navyblue", "violetred4")
isl <- c("North Island", "South Island", "Stewart Island")

mapColours <- data.frame(cols = I(cols), isl)

plot(nzSml84, col  = "grey", border = NA)
text(spatialCities, col = mapColours$cols[match(spatialCities$Island, mapColours$isl)],
label = spatialCities$City, cex = 1)
legend("topleft", text.col = mapColours$cols,
legend = mapColours$isl,
pch = 16)
}
