% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detailed_nz_islands}
\alias{detailed_nz_islands}
\title{detailed_nz_islands}
\format{
A SpatialPolygonsDataframe object with many
\describe{
  \item{name}{Name of the island}
  ...
}
}
\source{
\url{www.data.linz.govt.nz}
}
\usage{
detailed_nz_islands
}
\description{
A SpatialPolygonsDataframe of New Zealand in high resolutiono of a lot
of islands. Do not plot! But if you wish to extract islands manually
you can use this object.
}
\keyword{datasets}
